Add version to scheduler rpc API.
Part of blueprint versioned-rpc-apis.

One side effect of this change was that cinder.scheduler.api was removed
in favor of cinder.scheduler.rpcapi.  In this case, the api was just a
direct wrapper around rpc usage.  For other APIs, I've been following
the pattern that the rpcapi module provides the rpc client wrapper, and
if any other client-side logic is needed, that's where an api module is
used.

Change-Id: Ibd0292936f9afc77aeb5d040660bfa857861eed1