Remove cinder.log usage from cinder.rpc.
Part of blueprint common-rpc.

This patch removes the usage of cinder.log from cinder.rpc.  This is
necessary before cinder.rpc can be moved to openstack-common.  The rpc
logging was not using any of the extra features provided in cinder.log as
far as I can tell, so the changes are very simple.

This is the final dependency.  After this change, the rpc code
can be moved to openstack-common without any changes.

Change-Id: I17ca4b691bb8d4f18e618632a2381dd12d31852a